feat: 책장 페이지 클라이언트 컴포넌트 분리
